Added posibility to shop books

+++ src/main/java/com/example/demo/model/dto/request/BookRequest.java
@@ -2,7 +2,6 @@ package com.example.demo.model.dto.request;
 
 import com.example.demo.model.entity.Category;
 import lombok.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.Set;
 
@@ -18,7 +17,7 @@ public class BookRequest {
     private String author;
     @ToString.Exclude
     private byte[] image;
-    private Double price;
+    private Long price;
     private Long quantity;
     private Set<Category> categories;
 }
